requirements.txt
# here are the necessary libraries
discord.py == 1.6.0
spotipy == 2.17.1
lyricsgenius == 3.0.0
